requirements.txt

flask>=3.1.0
fastapi>=0.115.11
gunicorn>=23.0.0
openai>=1.65.4
faiss-cpu>=1.10.0
pydantic>=2.10.6
pypdf2>=3.0.1
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
python-multipart>=0.0.20
flask-wtf>=1.2.2
psutil>=7.0.0
pymupdf>=1.25.3
chromadb>=0.6.3
psycopg2-binary>=2.9.10
email-validator>=2.2.0
api>=0.0.7
numpy>=2.2.3
uvicorn>=0.34.0
requests
click
replit-object-storage
google-cloud-storage
